Add Java Maven index builder
Starting from `python build-maven-index.py https://repo.maven.apache.org/maven2/ ~/maven-central-index-working-dir/` for example, this will build `~/maven-central-index-working-dir/package-preferred-jars.txt`, a mapping from Java package names to the recommended order of JARs to use to satisfy that dependency. As much work as possible is done
incrementally, in order to allow incremental tweaks to the JAR selection algorithm and rapid rebuilds of the index.