Improve latest JAR version selection
This tries to account for the myriad ways people step outside the 'correct' format for Maven version numbers. We try to prefer the highest-numbered version while tolerating (but not preferring) arbitrary text suffixes, which can in practice be delimited by dots, dashes or nothing (e.g. "1.3beta4")

We also explicitly depreciate versions using very large numbers on the theory that they are most likely datestamps which we shouldn't mistake for a very-recent version.