Penalise packages that are provided by a groupId that has published much more recent packages (over a year newer)
The theory is these are likely to be obsolete. The penalty is only applied to other packages from the same groupId to try to distinguish a package that has been obsoleted and replaced, vs. one that is stable and hasn't been updated for a long time, but other groupIds sometimes publish copies or modified versions.